pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "classifieds-hub"
version = "0.1.0"
description = "Local classifieds aggregator for Aykhal and Udachny"
requires-python = ">=3.12"
dependencies = [
  "aiogram>=3.0,<4.0",
  "sqlalchemy>=2.0,<3.0",
  "aiosqlite>=0.20,<1.0",
  "apscheduler>=3.10,<4.0",
  "httpx>=0.27,<1.0",
  "pydantic-settings>=2.0,<3.0",
  "beautifulsoup4>=4.12,<5.0",
  "pillow>=10.0,<12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-cov>=5.0,<6.0",
  "ruff>=0.6.0",
  "mypy>=1.10",
  "pre-commit>=3.7",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]

[tool.mypy]
python_version = "3.12"
files = ["src"]
strict = false
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["apscheduler", "apscheduler.*", "apscheduler.schedulers.asyncio"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
